services/graphgen/requirements.txt
torch
numpy
scikit-learn
nltk
leidenalg
networkx
falkordb
python-dotenv
psycopg2-binary
pgvector
sentence-transformers
pydantic-settings
gliner
spacy
langchain-core
langchain-text-splitters
langchain-experimental
langchain-groq
fastapi
uvicorn
jellyfish
